First commit:
at the moment the binary tree is working but
the printing of the binary tree is crappy

We need now to have two function : one to encode,
the other to decode. The easiest way is to get a
list of bits from a text then encode it into binary
form for encoding.

To decode, one need to retrieve that list of bits
and decode sequentially from the dict-tree until
the total count of chars is matched